Implement CORS Configuration, Database Migrations, and Performance Enhancements
This commit introduces a comprehensive CORS configuration to restrict API access based on allowed origins, enhancing security for first-party frontends. Additionally, it includes new database migration scripts to add performance indexes for critical query patterns and to change the `amount` field in the `payment_setup` table from INTEGER to NUMERIC(12,2) to prevent data loss during transactions.

Furthermore, the commit refines the handling of large request bodies, reducing the limit from 50MB to 20MB to improve performance while still accommodating necessary payloads. The socket.io server configuration is updated to utilize the new CORS options, ensuring consistent behavior across the application.

Lastly, various controllers are updated to implement transaction locks for critical operations, preventing race conditions and ensuring data integrity during concurrent requests.

+++ src/controllers/admin/superAdmin/paymentSetupManagement.js
@@ -76,7 +76,7 @@ export const createPaymentSetup = TryCatchFunction(async (req, res) => {
   }
 
   // Validate amount is positive
-  const amountNum = parseInt(amount);
+  const amountNum = parseFloat(amount);
   if (isNaN(amountNum) || amountNum < 0) {
     throw new ErrorClass("amount must be a positive number", 400);
   }
@@ -153,7 +153,7 @@ export const updatePaymentSetup = TryCatchFunction(async (req, res) => {
   }
 
   if (amount !== undefined) {
-    const amountNum = parseInt(amount);
+    const amountNum = parseFloat(amount);
     if (isNaN(amountNum) || amountNum < 0) {
       throw new ErrorClass("amount must be a positive number", 400);
     }
